--- asm/bss_after_diskinit.s
@@ -1,6 +1,0 @@
-.include "macros.inc"
-
-.section .bss
-
-glabel D_801978D0
-.skip 12
