--- asm/bss_after_ostimer.s
@@ -1,6 +1,0 @@
-.include "macros.inc"
-
-.section .bss
-
-glabel D_801976A0
-.skip 432
