--- asm/bss_main_audio2.s
@@ -1,6 +1,0 @@
-.include "macros.inc"
-
-.section .bss
-
-glabel D_80194040
-.skip 16
