+++ courses/battle/block_fort/model.inc.c
@@ -1092,31 +1092,3 @@ mk64_Vtx d_course_block_fort_vertex[] = {
     {{  781,      83,     390}, {     0,     0}, {0x00, 0x01, 0x00, 0x00}},
 
 };
-
-/*
-int d_course_block_fort_packed[] = {
-    gsDPSetCombineMode(G_CC_SHADE, G_CC_SHADE),
-    gsSPTexture(qu016(1.52588e-05), qu016(1.52588e-05), 0, G_TX_RENDERTILE, G_OFF),
-    gsSPVertex(0x04000000, 3, 0),
-    gsSP2Triangles(0, 1, 2, 0, 0, 0, 0, 0),
-    gsSPTexture(qu016(0.999985), qu016(0.999985), 0, G_TX_RENDERTILE, G_ON),
-    gsDPSetCombineMode(G_CC_MODULATEIA, G_CC_MODULATEIA),
-    gsDPSetRenderMode(G_RM_AA_ZB_OPA_SURF, G_RM_AA_ZB_OPA_SURF2),
-    gsDPTileSync(),
-    gsDPSetTile(G_IM_FMT_RGBA, G_IM_SIZ_16b, 8, 0x0000, G_TX_RENDERTILE, 0, G_TX_NOMIRROR | G_TX_WRAP, 5, G_TX_NOLOD, G_TX_NOMIRROR | G_TX_WRAP, 5, G_TX_NOLOD),
-    gsDPSetTileSize(G_TX_RENDERTILE, 0, 0, qu102(31), qu102(31)),
-    gsDPSetTextureImage(G_IM_FMT_RGBA, G_IM_SIZ_16b, 1, 0x05000800),
-    gsDPTileSync(),
-    gsDPSetTile(G_IM_FMT_RGBA, G_IM_SIZ_16b, 0, 0x0000, G_TX_LOADTILE, 0, G_TX_NOMIRROR | G_TX_WRAP, G_TX_NOMASK, G_TX_NOLOD, G_TX_NOMIRROR | G_TX_WRAP, G_TX_NOMASK, G_TX_NOLOD),
-    gsDPLoadSync(),
-    gsDPLoadBlock(G_TX_LOADTILE, 0, 0, 1023, 256),
-    gsSPVertex(0x04000030, 20, 0),
-    gsSP2Triangles(0, 1, 2, 0, 0, 2, 3, 0),
-    gsSP2Triangles(4, 5, 6, 0, 4, 6, 7, 0),
-    gsSP2Triangles(8, 4, 7, 0, 9, 10, 11, 0),
-    gsSP2Triangles(9, 11, 12, 0, 13, 10, 9, 0),
-    gsSP2Triangles(11, 14, 12, 0, 15, 16, 17, 0),
-    gsSP2Triangles(15, 17, 18, 0, 5, 19, 6, 0),
-    gsSPEndDisplayList(),
-};
-*/
