+++ include/segment_symbols.h
@@ -1,6 +1,8 @@
 #ifndef SEGMENT_SYMBOLS_H
 #define SEGMENT_SYMBOLS_H
 
+extern u8 _mainSegNoloadEnd[];
+
 extern u8 _code_80280000SegmentStart[];
 extern u8 _code_80280000SegmentRomStart[];
 extern u8 _code_80280000SegmentRomEnd[];
