+++ src/actors/boat.inc.c
@@ -1,0 +1,3 @@
+void update_obj_paddle_wheel(struct PaddleWheelBoat *boat) {
+    boat->wheelRot += 0x38E;
+}
