+++ src/code_80005FD0.c
@@ -20,6 +20,11 @@
 #include "bomb_kart.h"
 #include "courses/all_course_data.h"
 
+
+//Collision D_80162E70;
+//s16 D_80162EB0; LIkely a float.
+
+
 // Strings, presented by google translate!
 // Note that these are EUC-JP encoded, see:
 // https://en.wikipedia.org/wiki/Extended_Unix_Code#EUC-JP
@@ -5789,6 +5794,7 @@ void func_800122D8(void) {
 //generated by m2c commit 8267401fa4ef7a38942dcca43353cc1bcc6efabc
 extern s16 D_80162EB0;
 
+// arg1 struct tbd
 void func_80012780(TrainCarStuff *trainCar, void *arg1, s16 arg2) {
     trainCar->position[0] = (f32) arg1->unk0;
     trainCar->position[1] = (f32) D_80162EB0;
